[K6.0] Make working RSS settings : recent and topics
